Add bucket deleting.
But you can't delete unless the keys are gone first.

The test originally failed because it didn't wait long enough for the
key to be deleted.  Turns out it took about four seconds for that to
happen.

I ended up reconfiguring the Riak cluster to have a delete_mode of
immediate, and that made the test run.

Chalk one up for not mocking every dependency.