Tune UTF-8 tests.
This came about because the last commit, while working on OSX, failed
under FreeBSD with all kinds of encoding errors.  I found the restify
author saying "yeah, none of that in URI's."

- Change tests to escape everything in the URI, per RFC3986.
- Add post version of getkey and delkey
- Put most UTF-8 test cases back into play.

At some point it might make sense to support IRI's, but for now it seems
a little safer shutting them down.

I'm not sure why the stuff in the malformed section still throws errors.