Test tweaks.
Use BUCKETLIST for seeding stubbed Riak buckets.

Use NODE_ENV to determine whether to stub or not.

Call it a stub, not a mock.  It's a little more descriptive.

Put in a hack to catch when the test server starts too quick to allow
the old server to drain away completely and restart it.  Mainly this
keeps test growl messages in good shape.

Only set a high timeout for the after routine when in the test
environment.  The stubbed riak doesn't need this.  It doesn't really
need the after_all in the first place, but until a speedup is needed,
I'm leaving it in.