Error and other refactoring.
Collect errors into single structure.  Makes it a little more readable.

Tiny refactoring tweaks in commands.

Added coverage for the unknown error case.

Threw in a case to cover a case where randomBytes returns a collision
with an existing bucket.  This is something that should really never
happen (at least within any reasonable time), and probably signals an
error with the crypto PRNG library or routine.