Stub riak by default, refactoring, cleanups.
Whenever the server is started in a non-production environment, stub out
the Riak client.  This allows for easy and quick manual testing of most
components.

Wrap the checker for a key/cert in a try.  Keep the rest out of a try,
just gate it based on whether the try worked.  It'd be nice if
LiveScript had a "try...then" construct.

Chop the telnet option timeout to 30ms to speed tests.  Hope that's
enough for production as well.

Wrap the riak stubber a little more, so random places don't have to
require the Riak module.

Stub out the request object for testing web site proxying.

The newbucket test is failing for some reason.  I'm on a plane, so I
can't set up a real Riak cluster and test this against it to tell what's
up.  Do this later!

Remove the contentTypeChecker middleware.

Extract the web proxy function.