Use a custom tgz for the Riak client
This is a hand-rolled version of the Riak client that includes a small
change that packs a huge performance improvement.  It caused the
duration of the regression tests to shrink from 3 minutes to 8 seconds.

With luck, the Basho upstream maintainer will merge in my pull request
(or rework it to make it fit what they prefer) and the performance
improvement will be widely available.  Until then, this hack.