Don't reject on media type parameters
Restify uses the negotiator module, which will fail to match media types
in the client's `accepts` header if there are parameters other than q
attached to them.  This results in 406 errors when something like
"Accepts: text/plain; charset=utf8" is sent.

Restify doesn't let you specify the charset as an acceptable parameter;
it strips that out of the server's list before checking.

This hack removes any charset specifiers from the accepts header.  It
should probably be removed at a later time, after the negotiator module
is fixed (see https://github.com/jshttp/negotiator/issues/35)

I found this as swagger-ui's requests have charsets appended to the
accepts header, so it was failing with 406 errors.