More swagger
This adds a 'group' property to commands, which is a semantic grouping
of command types.  Currently, there's "buckets," "keys," and "search."
These are used to section the API tester and docs.

Removed the .ls extension from the main require in kvs.io to accommodate
pre-compiling ls files to .js.

Move is_prod to package level in api.ls.  Shouldn't change during
runtime.

Refactor the swagger operation command.  Turns out there's a bit more
subtlety in how this stuff is done.

Set host to be localhost:8080 for testing.  This'll change pretty soon.

Remove last try at swaggerification.

Only honor SIGUSR2 in production.  This signal is used in production to
release and reload logs so data isn't lost during the log rotation
process.  However, nodemon uses USR2 to signal the child to restart when
the code changes.

Move swagger-ui to just swagger.