Tighten up error handling, minor refactoring
I've got each command reporting the errors it can send back.  Each could
potentially also send back something that'll get translated to a 500, so
show that also.

Host is kvs.io on production, localhost everywhere else.

Moved the API init to after the server is listening.  This lets the code
get a port number properly.

refactor is_prod for cli.

Change doc -> summary to comply with Swagger rules more easily.

Trim timeouts in test/dev.  Not sure if these are as low as are likely,
will tune as time goes on.

Removed an error that'