Add the sleep command.  Don't prompt until last command finishes.
This is primarily so I can produce output after the connection closes.

It can also keep the connection open for a long time.  Eventually, the
signal that someone has dropped a connection should be caught so that a
zillion sleeps can't be issued to keep connections hanging.

This also makes the CLI synchronous, which is something I had assumed
when writing on the tests.  Hopefully it means the tests will end up
more stable.