Step three for long-polling callbacks
This final form accommodates multiple kvsio servers, each with multiple
processes.

When a listen request comes in, a callback is registered on a bucket
with a URL hitting the IP address of this server.  Instead of
responding, the handler returns to allow the event loop to continue.
This keeps the socket open and the listener hanging.

When an operation is triggered and a callback is fired, a Restify server
on this host receives it.  It then pulls a process ID out of the
callback URL and sends a message via ZMQ to that process.  The process
receives that event, finds the original listener, and completes the
callback.

It's not completely finished yet, but it's real close.