Refactor CLI stuff
The CLI stuff wasn't re-entrant.  This makes it so.  Also added a couple
tests for multiple simultaneous clients.  Means the CLI now depends on
the API to operate to transfer messages for listeners.

There is some problem when upping the number of clients on the "whole
lotta listening" test.  More than 5 gives some socket errors.  I'm
hoping this has to do with per-process socket limits on my Mac.  It
should be investigated at some point.